backend/requirements.txt
asyncpg==0.21.0
click==7.1.2
databases==0.4.1
fastapi==0.63.0
freeze==3.0
h11==0.11.0
loguru==0.5.3
psycopg2-binary
pydantic==1.7.3
python-dotenv==0.15.0
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.2